Address ADIv6 access ports through their debug-space bases.
ADIv6 uses aligned AP base addresses and 4 KiB register windows instead
of ADIv5 APSEL indices. Reusing an index would select the wrong
register.

Keep those selectors distinct, check them against the connected DP and
its advertised address width, and select both address words when needed.
Immediate AP reads and writes retain posted completion and existing
recovery. Queued AP operations and MEM-AP acquisition remain ADIv5-only.

+++ dap/ap_integration_test.go
@@ -23,7 +23,7 @@ func TestAccessAPOverFTDI(t *testing.T) {
 	defer cancel()
 
 	dp := openHardwareDebugPort(t, ctx)
-	const hardwareAPCSW = uint8(0x00)
+	const hardwareAPCSW = uint16(0x00)
 	var (
 		savedCSW uint32
 		saved    bool
